% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_equivalent_moments.R
\name{Compute_equivalent_moments}
\alias{Compute_equivalent_moments}
\title{Computing equivalent sample moments given the parameters of a ghMST distribution}
\usage{
Compute_equivalent_moments(parameters)
}
\arguments{
\item{parameters}{List of fitted parameters, including location vector, skewness vector, scatter matrix, and the degree of freedom..}
}
\value{
A list containing the following elements:
\item{\code{mu}}{Mean vector.}
\item{\code{Sgm}}{Covariance matrix.}
\item{\code{Phi_mat}}{Co-skewness matrix.}
\item{\code{Psi_mat}}{Co-kurtosis matrix.}
\item{\code{Phi}}{Co-skewness matrix in vector form (collecting only the unique elements).}
\item{\code{Psi}}{Co-kurtosis matrix in vector form (collecting only the unique elements).}
\item{\code{Phi_shred}}{Partition on \code{Phi} (see reference).}
\item{\code{Psi_shred}}{Partition on \code{Psi} (see reference).}
}
\description{
Computing equivalent sample moments given the parameters of a ghMST distribution
}
\examples{
library(ParametricHighOrderPortfolios)
data(X50)

parameters <- fit_ghMST(X50)
# decide moment weights
equivalent_sample_moments <- Compute_equivalent_moments(parameters)

}
\references{
X. Wang, R. Zhou, J. Ying, and D. P. Palomar, "Efficient and Scalable High-Order Portfolios Design via Parametric Skew-t Distribution,"
Available in arXiv, 2022. <https://arxiv.org/pdf/2206.02412.pdf>.
R. Zhou and D. P. Palomar, "Solving High-Order Portfolios via Successive Convex Approximation Algorithms,"
in \emph{IEEE Transactions on Signal Processing}, vol. 69, pp. 892-904, 2021.
<doi:10.1109/TSP.2021.3051369>.
}
\author{
Xiwen Wang, Daniel P. Palomar
}
