% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nloptr_non_parametric.R
\name{solve_MVSK_nloptr_non_parametric}
\alias{solve_MVSK_nloptr_non_parametric}
\title{Design MVSK portfolio via nloptr based on sample moments}
\usage{
solve_MVSK_nloptr_non_parametric(lambda, X_moments, w0)
}
\arguments{
\item{lambda}{Numerical vector of length 4 indicating the weights of first four moments.}

\item{X_moments}{List of moment parameters, see \code{\link{estimate_moments}}.}

\item{w0}{Numerical vector indicating the initial value of portfolio weights.}
}
\value{
A list containing the following elements:
\item{\code{w_optimal}}{Optimal portfolio vector.}
\item{\code{obj_optimal}}{Optimal objective value.}
\item{\code{cpu_time}}{Computational time.}
}
\description{
Design MVSK portfolio via nloptr based on sample moments
}
\author{
Xiwen Wang and Daniel P. Palomar
}
