% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nloptr_skewt.R
\name{solve_MVSK_nloptr_skewt}
\alias{solve_MVSK_nloptr_skewt}
\title{Design MVSK portfolio via nloptr based on the parameters of generalized hyperbolic skew-t distribution}
\usage{
solve_MVSK_nloptr_skewt(w0, lambda, parameters)
}
\arguments{
\item{w0}{Numerical vector indicating the initial value of portfolio weights.}

\item{lambda}{Numerical vector of length 4 indicating the weights of first four moments.}

\item{parameters}{List of fitted parameters, including location vector, skewness vector, scatter matrix, and the degree of freedom..}
}
\value{
A list containing the following elements:
\item{\code{w_optimal}}{Optimal portfolio vector.}
\item{\code{obj_optimal}}{Optimal objective value.}
\item{\code{cpu_time}}{Computational time.}
}
\description{
Design MVSK portfolio via nloptr based on the parameters of generalized hyperbolic skew-t distribution:
}
\author{
Xiwen Wang and Daniel P. Palomar
}
