% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSK_non_parametric.R
\name{solve_MVSK_non_parametric}
\alias{solve_MVSK_non_parametric}
\title{Solve high-order portfolio without shorting based on non-parametric representation of the high-order moments}
\usage{
solve_MVSK_non_parametric(
  lmd,
  X_moments,
  w_init = rep(1/length(X_moments$mu), length(X_moments$mu)),
  method = c("Q-MVSK", "L-MVSK", "DC", "PGD", "RFPA", "SQUAREM"),
  initial_eta = 10,
  beta = 0.5,
  tau = 1e+06,
  tau_w = 0,
  gamma = 1,
  zeta = 1e-08,
  maxiter = 1000,
  ftol = 1e-06,
  wtol = 1e-06,
  stopval = -Inf
)
}
\arguments{
\item{lmd}{Numerical vector of length 4 indicating the weights of first four moments.}

\item{X_moments}{List of moment parameters, see \code{\link{estimate_moments}}.}

\item{w_init}{Numerical vector indicating the initial value of portfolio weights.}

\item{method}{String indicating the algorithm method, must be one of: "Q-MVSK", "L-MVSK", "DC", "PGD", "RFPA", "SQUAREM".}

\item{initial_eta}{initial eta for the PGD and RFPA methods}

\item{beta}{scaling factor for the PGD and RFPA methods}

\item{tau}{hyper-parameter tau for the PGD and SQUAREM methods}

\item{tau_w}{Number (>= 0) guaranteeing the strong convexity of approximating function.}

\item{gamma}{Number (0 < gamma <= 1) indicating the initial value of gamma for the Q-MVSK method.}

\item{zeta}{Number (0 < zeta < 1) indicating the diminishing parameter of gamma for the Q-MVSK method.}

\item{maxiter}{Positive integer setting the maximum iteration.}

\item{ftol}{Positive number setting the convergence criterion of function objective.}

\item{wtol}{Positive number setting the convergence criterion of portfolio weights.}

\item{stopval}{Number setting the stop value of objective.}
}
\value{
A list containing the following elements:
\item{\code{w}}{Optimal portfolio vector.}
\item{\code{cpu_time_vs_iterations}}{Time usage over iterations.}
\item{\code{objfun_vs_iterations}}{Objective function over iterations.}
\item{\code{iterations}}{Iterations index.}
\item{\code{convergence}}{Boolean flag to indicate whether or not the optimization converged.}
\item{\code{moments}}{Moments of portfolio return at optimal portfolio weights.}
}
\description{
Solve high-order portfolio without shorting based on non-parametric representation of the high-order moments
(i.e., mean, variance, skewness, and kurtosis):
\preformatted{
  minimize     - lmd1*(w'*mu) + lmd2*(w'*Sigma*w)
               - lmd3*(w'*Phi*w*w) + lmd4*(w'*Psi*w*w*w)
  subject to   w>=0, sum(w) == 1,
}
}
\examples{

library(ParametricHighOrderPortfolios)
data(X50)

# estimate moments
X_moments <- estimate_moments(X50[, 1:10])

# decide moment weights
lmd <- c(1,4,10,20)

# portfolio optimization
sol <- solve_MVSK_non_parametric(lmd, X_moments, method = "Q-MVSK")


}
\references{
R. Zhou and D. P. Palomar, "Solving High-Order Portfolios via Successive Convex Approximation Algorithms,"
in \emph{IEEE Transactions on Signal Processing}, vol. 69, pp. 892-904, 2021.
<doi:10.1109/TSP.2021.3051369>.

X. Wang, R. Zhou, J. Ying, and D. P. Palomar, "Efficient and Scalable High-Order Portfolios Design via Parametric Skew-t Distribution,"
Available in arXiv, 2022. <https://arxiv.org/pdf/2206.02412.pdf>.
}
\author{
Xiwen, Rui Zhou and Daniel P. Palomar
}
